New utils file for checking for corresponding bids files
